shelves: Refactor shelf feature
