+++ client/src/features/shelf/index.ts
@@ -1,4 +1,8 @@
-// UI
-export { default as CreateShelfDialog } from './ui/create-shelf/create-shelf-dialog';
+// UI -> Create Shelf
+export { default as CreateShelfDialog } from './ui/create-shelf';
+export { default as CreateShelfDialogContent } from './ui/create-shelf/content';
 export { default as CreateShelfBtn } from './ui/create-shelf/create-shelf-btn';
 export { default as CreateShelfWidget } from './ui/create-shelf/create-shelf-widget';
+
+// UI -> Delete Shelf
+// ...
