+++ client/src/features/shelf/ui/create-shelf/index.tsx
@@ -1,0 +1,23 @@
+// Basics
+import { ReactNode } from 'react';
+
+// Shadcn / Tailwind
+import { Dialog, DialogTrigger } from "@/shared/ui/dialog";
+
+// Insides
+import CreateShelfDialogContent from './content';
+
+// Common modal window for shelf creation
+const CreateShelfDialog = ({ children }: { children: ReactNode }) => (
+    <Dialog>
+
+        <DialogTrigger asChild>
+            {children}
+        </DialogTrigger>
+
+        <CreateShelfDialogContent />
+
+    </Dialog>
+)
+
+export default CreateShelfDialog;
